requirements.txt
appdirs==1.4.3
asgiref==3.2.5
boto3==1.12.19
botocore==1.15.19
cfgv==3.1.0
distlib==0.3.0
Django==3.0.4
django-extensions==2.2.8
django-model-utils==4.0.0
django-storages==1.9.1
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.11
importlib-metadata==1.5.0
jmespath==0.9.5
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.3.5
Pillow==7.0.0
pre-commit==2.2.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.10
Werkzeug==1.0.0
zipp==3.1.0
